# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/banners": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get list of banners",
                "operationId": "getBannerList",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "image": "https://www.bhdstar.vn/wp-content/uploads/2024/07/Banner-Web-BHD_Milk.jpg",
                                                    "type": "main",
                                                    "status": true
                                                },
                                                {
                                                    "id": 2,
                                                    "image": "https://www.bhdstar.vn/wp-content/uploads/2024/07/Banner-Web-BHD_Milk2.jpg",
                                                    "type": "sub",
                                                    "status": true
                                                }
                                            ]
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Lấy danh sách banner thành công"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },

            "post": {
                "tags": [
                    "Banners"
                ],
                "summary": "Create a new banner",
                "operationId": "createBanner",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Tải ảnh banner lên server."
                                    }
                                },
                                "required": ["image"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tải ảnh lên thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "uploaded-file.jpg"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Tải ảnh lên thành công."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dữ liệu đầu vào không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "file": {
                                                    "type": "array",
                                                    "example": [
                                                        "Bạn chưa chọn file ảnh để tải lên.",
                                                        "Vui lòng chọn file đúng định dạng ảnh như png, jpeg, jpg, jfif.",
                                                        "Kích thước ảnh quá lớn, vui lòng chọn ảnh < 5MB."
                                                    ]
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/banners/{id}": {
            "get": {
                "tags": [
                    "Banners"
                ],
                "summary": "Get banner by ID",
                "operationId": "getBannerById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A single banner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "image_url": "https://www.bhdstar.vn/wp-content/uploads/2024/07/Banner-Web-BHD_Milk.jpg",
                                                "type": "main",
                                                "status": true
                                            }
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Lấy banner thành công!"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Voucher not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Banners"
                ],
                "summary": "Delete a banner",
                "operationId": "deleteBanner",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "ID của banner cần xóa"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Banner đã được xóa thành công.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": null,
                                            "example": null
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Banner đã được xóa thành công."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy banner",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "status": {
                                                    "type": "boolean",
                                                    "example": false
                                                },
                                                "id": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Không tìm thấy ID của banner."
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}